added directory traversal option to remote_delete()

+++ sftpclone/sftpclone.py
@@ -400,10 +400,13 @@ class SFTPClone(object):
         """Remove the remote directory node."""
         # If it's a directory, then delete content and directory
         if S_ISDIR(r_st.st_mode):
-            for item in self.sftp.listdir_attr(remote_path):
-                full_path = path_join(remote_path, item.filename)
-                self.remote_delete(full_path, item)
-            self.sftp.rmdir(remote_path)
+            if self.traverse_remote_directories:
+                for item in self.sftp.listdir_attr(remote_path):
+                    full_path = path_join(remote_path, item.filename)
+                    self.remote_delete(full_path, item)
+                    self.sftp.rmdir(remote_path)
+            else:
+                self.logger.info("skipping %s as traverse_remote_directories is off", remote_path)
 
         # Or simply delete files
         else:
